cleanup, moving stuff to common, nuke utils
